Refactor Speedreader + DeAmp throttle/loader
