+++ components/sniffer/DEPS
@@ -1,7 +1,6 @@
 include_rules = [
   "+content/public/browser",
-  "+components/sniffer",
   "+services/network/public/cpp",
   "+services/network/public/mojom",
-  "+third_party/re2",
+  "+third_party/blink/public/common/loader",
 ]
