+++ components/speedreader/BUILD.gn
@@ -45,6 +45,7 @@ static_library("speedreader") {
     "//brave/components/brave_component_updater/browser",
     "//brave/components/resources",
     "//brave/components/resources:static_resources_grit",
+    "//brave/components/sniffer",
     "//brave/components/weekly_storage",
     "//components/content_settings/core/browser",
     "//components/content_settings/core/common",
