+++ components/speedreader/speedreader_throttle.cc
@@ -20,6 +20,9 @@
 #include "services/network/public/mojom/url_response_head.mojom.h"
 
 namespace speedreader {
+
+// SpeedReaderThrottle::~SpeedReaderThrottle() = default;
+
 // static
 std::unique_ptr<SpeedReaderThrottle>
 SpeedReaderThrottle::MaybeCreateThrottleFor(
@@ -40,9 +43,9 @@ SpeedReaderThrottle::SpeedReaderThrottle(
     SpeedreaderRewriterService* rewriter_service,
     base::WeakPtr<SpeedreaderResultDelegate> result_delegate,
     scoped_refptr<base::SingleThreadTaskRunner> task_runner)
-    : rewriter_service_(rewriter_service),
-      result_delegate_(result_delegate),
-      task_runner_(std::move(task_runner)) {}
+    : task_runner_(task_runner),
+      rewriter_service_(rewriter_service),
+      result_delegate_(result_delegate) {}
 
 SpeedReaderThrottle::~SpeedReaderThrottle() = default;
 
@@ -51,8 +54,6 @@ void SpeedReaderThrottle::WillProcessResponse(
     network::mojom::URLResponseHead* response_head,
     bool* defer) {
   VLOG(2) << "Speedreader throttling: " << response_url;
-  // Pause the response until Speedreader has done its job.
-  *defer = true;
 
   mojo::PendingRemote<network::mojom::URLLoader> new_remote;
   mojo::PendingReceiver<network::mojom::URLLoaderClient> new_receiver;
@@ -64,14 +65,10 @@ void SpeedReaderThrottle::WillProcessResponse(
       SpeedReaderURLLoader::CreateLoader(weak_factory_.GetWeakPtr(),
                                          result_delegate_, response_url,
                                          task_runner_, rewriter_service_);
-  delegate_->InterceptResponse(std::move(new_remote), std::move(new_receiver),
-                               &source_loader, &source_client_receiver, &body);
-  speedreader_loader->Start(std::move(source_loader),
-                            std::move(source_client_receiver), std::move(body));
-}
-
-void SpeedReaderThrottle::Resume() {
-  delegate_->Resume();
+  SnifferThrottle::InterceptAndStartLoader(
+      std::move(source_loader), std::move(source_client_receiver),
+      std::move(new_remote), std::move(new_receiver), speedreader_loader,
+      defer);
 }
 
 }  // namespace speedreader
