+++ components/speedreader/speedreader_url_loader.h
@@ -15,6 +15,7 @@
 #include "base/memory/ref_counted.h"
 #include "base/memory/weak_ptr.h"
 #include "base/strings/string_piece.h"
+#include "brave/components/sniffer/sniffer_url_loader.h"
 #include "brave/components/speedreader/speedreader_result_delegate.h"
 #include "mojo/public/cpp/bindings/pending_receiver.h"
 #include "mojo/public/cpp/bindings/pending_remote.h"
@@ -34,6 +35,8 @@ class SpeedreaderRewriterService;
 
 // Loads the whole response body and tries to Speedreader-distill it.
 // Cargoculted from |`SniffingURLLoader|.
+// Note that common functionality between this class and DeAmp has
+// been moved to component/sniffer
 //
 // This loader has five states:
 // kWaitForBody: The initial state until the body is received (=
@@ -53,34 +56,23 @@ class SpeedreaderRewriterService;
 // kAborted: Unexpected behavior happens. Watchers, pipes and the binding from
 //           the source loader to |this| are stopped. All incoming messages from
 //           the destination (through network::mojom::URLLoader) are ignored in
-class SpeedReaderURLLoader : public network::mojom::URLLoaderClient,
-                             public network::mojom::URLLoader {
+class SpeedReaderURLLoader : public sniffer::SnifferURLLoader {
  public:
   ~SpeedReaderURLLoader() override;
-
-  SpeedReaderURLLoader(const SpeedReaderURLLoader&) = delete;
-  SpeedReaderURLLoader& operator=(const SpeedReaderURLLoader&) = delete;
-
-  // Start waiting for the body.
-  void Start(
-      mojo::PendingRemote<network::mojom::URLLoader> source_url_loader_remote,
-      mojo::PendingReceiver<network::mojom::URLLoaderClient>
-          source_url_client_receiver,
-      mojo::ScopedDataPipeConsumerHandle body);
 
   // mojo::PendingRemote<network::mojom::URLLoader> controls the lifetime of the
   // loader.
   static std::tuple<mojo::PendingRemote<network::mojom::URLLoader>,
                     mojo::PendingReceiver<network::mojom::URLLoaderClient>,
                     SpeedReaderURLLoader*>
-  CreateLoader(base::WeakPtr<SpeedReaderThrottle> throttle,
+  CreateLoader(base::WeakPtr<sniffer::SnifferThrottle> throttle,
                base::WeakPtr<SpeedreaderResultDelegate> delegate,
                const GURL& response_url,
                scoped_refptr<base::SingleThreadTaskRunner> task_runner,
                SpeedreaderRewriterService* rewriter_service);
 
  private:
-  SpeedReaderURLLoader(base::WeakPtr<SpeedReaderThrottle> throttle,
+  SpeedReaderURLLoader(base::WeakPtr<sniffer::SnifferThrottle> throttle,
                        base::WeakPtr<SpeedreaderResultDelegate> delegate,
                        const GURL& response_url,
                        mojo::PendingRemote<network::mojom::URLLoaderClient>
@@ -88,77 +80,15 @@ class SpeedReaderURLLoader : public network::mojom::URLLoaderClient,
                        scoped_refptr<base::SingleThreadTaskRunner> task_runner,
                        SpeedreaderRewriterService* rewriter_service);
 
-  // network::mojom::URLLoaderClient implementation (called from the source of
-  // the response):
-  void OnReceiveEarlyHints(network::mojom::EarlyHintsPtr early_hints) override;
-  void OnReceiveResponse(network::mojom::URLResponseHeadPtr response_head,
-                         mojo::ScopedDataPipeConsumerHandle body) override;
-  void OnReceiveRedirect(
-      const net::RedirectInfo& redirect_info,
-      network::mojom::URLResponseHeadPtr response_head) override;
-  void OnUploadProgress(int64_t current_position,
-                        int64_t total_size,
-                        OnUploadProgressCallback ack_callback) override;
-  void OnReceiveCachedMetadata(mojo_base::BigBuffer data) override;
-  void OnTransferSizeUpdated(int32_t transfer_size_diff) override;
-  void OnStartLoadingResponseBody(
-      mojo::ScopedDataPipeConsumerHandle body) override;
-  void OnComplete(const network::URLLoaderCompletionStatus& status) override;
-
-  // network::mojom::URLLoader implementation (called from the destination of
-  // the response):
-  void FollowRedirect(
-      const std::vector<std::string>& removed_headers,
-      const net::HttpRequestHeaders& modified_headers,
-      const net::HttpRequestHeaders& modified_cors_exempt_headers,
-      const absl::optional<GURL>& new_url) override;
-  void SetPriority(net::RequestPriority priority,
-                   int32_t intra_priority_value) override;
-  void PauseReadingBodyFromNet() override;
-  void ResumeReadingBodyFromNet() override;
-
-  void OnBodyReadable(MojoResult);
-  void OnBodyWritable(MojoResult);
+  void OnBodyReadable(MojoResult) override;
+  void OnBodyWritable(MojoResult) override;
   void MaybeLaunchSpeedreader();
 
-  // Gets either distilled or untouched body.
-  void CompleteLoading(std::string body);
-  void CompleteSending();
-  void SendReceivedBodyToClient();
-
-  void Abort();
-
-  base::WeakPtr<SpeedReaderThrottle> throttle_;
+  void CompleteSending() override;
   base::WeakPtr<SpeedreaderResultDelegate> delegate_;
-
-  mojo::Receiver<network::mojom::URLLoaderClient> source_url_client_receiver_{
-      this};
-  mojo::Remote<network::mojom::URLLoader> source_url_loader_;
-  mojo::Remote<network::mojom::URLLoaderClient> destination_url_loader_client_;
-
-  GURL response_url_;
-
-  scoped_refptr<base::SingleThreadTaskRunner> task_runner_;
-
-  enum class State { kWaitForBody, kLoading, kSending, kCompleted, kAborted };
-  State state_ = State::kWaitForBody;
-
-  // Set if OnComplete() is called during distilling.
-  absl::optional<network::URLLoaderCompletionStatus> complete_status_;
-
-  // Note that this could be replaced by a distilled version.
-  std::string buffered_body_;
-  size_t bytes_remaining_in_buffer_;
-
-  mojo::ScopedDataPipeConsumerHandle body_consumer_handle_;
-  mojo::ScopedDataPipeProducerHandle body_producer_handle_;
-  mojo::SimpleWatcher body_consumer_watcher_;
-  mojo::SimpleWatcher body_producer_watcher_;
 
   // Not Owned
   raw_ptr<SpeedreaderRewriterService> rewriter_service_ = nullptr;
-
-  base::WeakPtrFactory<SpeedReaderURLLoader> weak_factory_{this};
 };
 
 }  // namespace speedreader
